fix: refactoring, add features ssr in mod server.rs
