+++ app/src/common.rs
@@ -1,1 +1,2 @@
 pub(crate) mod dto;
+pub(crate) mod handlers;
