+++ app/src/server.rs
@@ -1,6 +1,5 @@
 pub(crate) mod models;
 pub(crate) mod services;
-#[cfg(feature = "ssr")]
-pub mod core;
+pub mod contexts;
 
 pub mod config;
